streamline types
