+++ lib/llm/OpenAIClient.ts
@@ -1,9 +1,10 @@
 import OpenAI, { ClientOptions } from "openai";
 import { zodResponseFormat } from "openai/helpers/zod";
-import { LLMClient, ChatCompletionOptions } from "./LLMClient";
-import { LLMCache } from "../cache/LLMCache";
-import { LogLine, AvailableModel } from "../types";
 import { ChatCompletionCreateParamsNonStreaming } from "openai/resources/chat";
+import { LogLine } from "../../types/log";
+import { AvailableModel } from "../../types/model";
+import { LLMCache } from "../cache/LLMCache";
+import { ChatCompletionOptions, LLMClient } from "./LLMClient";
 
 export class OpenAIClient extends LLMClient {
   private client: OpenAI;
