+++ lib/utils.ts
@@ -1,5 +1,5 @@
 import crypto from "crypto";
-import { LogLine } from "./types";
+import { LogLine } from "../types/log";
 
 export function generateId(operation: string) {
   return crypto.createHash("sha256").update(operation).digest("hex");
