+++ types/browser.ts
@@ -1,0 +1,8 @@
+import { Browser, BrowserContext } from "@playwright/test";
+
+export interface BrowserResult {
+  browser?: Browser;
+  context: BrowserContext;
+  debugUrl?: string;
+  sessionUrl?: string;
+}
