+++ types/model.ts
@@ -3,34 +3,6 @@ import { Tool as AnthropicTool } from "@anthropic-ai/sdk/resources";
 import type { ClientOptions as OpenAIClientOptions } from "openai";
 import { ChatCompletionTool as OpenAITool } from "openai/resources";
 
-export class PlaywrightCommandException extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "PlaywrightCommandException";
-  }
-}
-
-export class PlaywrightCommandMethodNotSupportedException extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "PlaywrightCommandMethodNotSupportedException";
-  }
-}
-
-export type LogLine = {
-  id?: string;
-  category?: string;
-  message: string;
-  level?: 0 | 1 | 2;
-  timestamp?: string;
-  auxiliary?: {
-    [key: string]: {
-      value: string;
-      type: "object" | "string" | "html" | "integer" | "float" | "boolean";
-    };
-  };
-};
-
 export type AvailableModel =
   | "gpt-4o"
   | "gpt-4o-mini"
