+++ types/playwright.ts
@@ -1,0 +1,13 @@
+export class PlaywrightCommandException extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "PlaywrightCommandException";
+  }
+}
+
+export class PlaywrightCommandMethodNotSupportedException extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "PlaywrightCommandMethodNotSupportedException";
+  }
+}
